refactor(core): Move params interfaces
