+++ packages/core/interfaces/context/index.ts
@@ -1,3 +1,4 @@
 export * from "./arguments-host.interface";
 export * from "./creator.interface";
 export * from "./execution.interface";
+export * from "./params.interface";
